refactor: repackage redis config and service
