+++ module-api/src/main/java/com/codingbottle/common/redis/config/LikeRedisConfig.java
@@ -1,6 +1,6 @@
-package com.codingbottle.common.config;
+package com.codingbottle.common.redis.config;
 
-import org.springframework.beans.factory.annotation.Qualifier;
+import com.codingbottle.common.config.RedisConfig;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
@@ -8,7 +8,7 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 
 @Configuration
-@EnableRedisRepositories
+@EnableRedisRepositories(redisTemplateRef = "likesRedisTemplate")
 public class LikeRedisConfig extends RedisConfig {
     @Bean
     public RedisConnectionFactory deviceControlRedisConnectionFactory() {
@@ -16,7 +16,6 @@ public class LikeRedisConfig extends RedisConfig {
     }
 
     @Bean
-    @Qualifier("likesRedisTemplate")
     public RedisTemplate<Long, Long> likesRedisTemplate() {
         RedisTemplate<Long, Long>  template = new RedisTemplate<>();
         template.setConnectionFactory(deviceControlRedisConnectionFactory());
